+++ simulator/simulator/analyzer.py
@@ -8,7 +8,7 @@ Created: 10/9/24
 # ROS MODULES
 import rclpy
 from rclpy.node import Node
-from std_msgs.msg import UInt32
+from std_msgs.msg import UInt32, Float64
 from nav_msgs.msg import Odometry
 
 # FILE WRITING
@@ -42,13 +42,13 @@ class Analyzer(Node):
 
                 # SUBSCRIBERS - INPUTS (CONTROL VARIABLES)
                 self.left_pwm_sub = self.create_subscription(
-                        UInt32,
+                        Float64,
                         "/steering_left/percentage",
                         self.left_pwm_callback,
                         10
                 )
                 self.right_pwm_sub = self.create_subscription(
-                        UInt32,
+                        Float64,
                         "/steering_right/percentage",
                         self.right_pwm_callback,
                         10
